# perf(screenshot): use sampling for `is-white`