… packages/screenshot/test/fixtures/no-white-5k.jpg
… packages/screenshot/test/fixtures/no-white-5k.png
… packages/screenshot/test/fixtures/white-5k.jpg
… packages/screenshot/test/fixtures/white-5k.png